wbsite update
added id attributes to main page and signup page. got rid of some of the pictures from the template that we weren't using. changed the background to a more solid style to see what people thought and added signup fail and kong grow fail htmls that will be the pages displayed when log in fails or a user name is already taken.